requirements.txt
Adafruit-PlatformDetect==3.59.0
adafruit-python-shell==1.8.0
args==0.1.0
asn1crypto==0.24.0
autopep8==2.0.4
certifi==2018.8.24
chardet==3.0.4
click==8.1.7
clint==0.5.1
colorzero==1.1
cryptography==2.6.1
docutils==0.20.1
entrypoints==0.3
evdev==1.1.2
gpiozero==1.6.2
idna==2.6
importlib-metadata==6.7.0
keyring==17.1.1
keyrings.alt==3.1.1
Kivy==2.3.0
Kivy-examples==2.3.0
Kivy-Garden==0.1.5
paho-mqtt==1.6.1
pigpio==1.78
Pillow==9.5.0
psutil==5.9.8
pybleno==0.11
pycodestyle==2.10.0
pycrypto==2.6.1
pygame==2.5.2
pygments==2.17.2
PyGObject==3.30.4
python-apt==1.8.4.3
pyxdg==0.25
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
spidev==3.5
ssh-import-id==5.7
tomli==2.0.1
typing-extensions==4.7.1
urllib3==1.24.1
zipp==3.15.0
